## USER PROFILE - User Profile, Video Owner, User's Videos, Bugfix

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from '../models/Video';
+import User from '../models/User';
 
 // export const home = (req, res) => {
 //   Video.find({})
@@ -19,8 +20,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
-  // console.log(video);
+  const video = await Video.findById(id).populate('owner');
 
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found.' });
@@ -30,21 +30,33 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
 
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   res.render('edit', { pageTitle: `Edit : ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
 
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -59,16 +71,23 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
+
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
-
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect('/');
   } catch (error) {
     return res
@@ -79,9 +98,18 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
-  // delete video
-  await Video.findByIdAndDelete(id);
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
 
+  if (!video) {
+    return res.render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
+  await Video.findByIdAndDelete(id);
   return res.redirect('/');
 };
 
@@ -96,6 +124,5 @@ export const search = async (req, res) => {
       },
     }).sort({ createdAt: 'desc' });
   }
-
   return res.render('search', { pageTitle: 'Search', videos });
 };
